Improve and simplify the unit tests
